+++ cassava-megaparsec.cabal
@@ -1,5 +1,5 @@
 name:                 cassava-megaparsec
-version:              2.0.3
+version:              2.1.0
 cabal-version:        1.18
 tested-with:          GHC==7.10.3, GHC==8.0.2, GHC==8.2.2, GHC==8.4.3, GHC==8.6.1
 license:              MIT
@@ -32,6 +32,7 @@ library
                     , unordered-containers >= 0.2.7 && < 0.3
                     , vector           >= 0.11  && < 0.13
   exposed-modules:    Data.Csv.Parser.Megaparsec
+                    , Data.Csv.Parser.Megaparsec.Internals
   if flag(dev)
     ghc-options:      -Wall -Werror
   else
